requirements.txt
Flask
python-docx
gunicorn
PyJWT
bcrypt
sqlite-utils
